requirements.txt
html5lib
requests
beautifulsoup4
xlsxwriter
PyYAML